requirements.txt
streamlit 
google-generativeai  
python-dotenv 
langchain 
PyPDF2  
faiss-cpu  
langchain_google_genai 
pdf2image
poppler-utils
streamlit_lottie
spacy
langchain-community
beautifulsoup4
numpy
pandas
Requests
requests_html

streamlit-webrtc
aiortc
flask
flask-socketio
streamlit_chat

streamlit-on-Hover-tabs
httpx-oauth
firebase-admin
streamlit-authenticator
youtube-transcript-api
genai